llm_balance_demo/requirements.txt
fastapi~=0.85.0
uvicorn
openai~=1.82.0
python-dotenv
pydantic~=1.10.2